ci: Add .gitlab-ci.yml based on gitlab template
Use docker image with prebuild dependencies.
This speeds up the tests.
It has the php dependencies installed
and nextcloud server installed in /srv/nextcloud/server

Test currently runs as two jobs:
 * one for integration tests (currently failing)
 * one for unit tests (none for now)

The tests are failing the same way they do for me locally.
Even if this is fixed they may continue to fail
because the database connection is not setup yet.

The git checkout is symlinked into `/srv/nextcloud/server/apps`.
This way the app will be registered in nextcloud.

tests/bootstrap.php will look for the nextcloud source in
`$NEXTCLOUD_SOURCE`
and fallback to the relative path of nextcloud
as seen from an install inside the apps directory.

Includes setup and config for code coverage.
Currently it is at 0% because tests are missing or failing.