Use the markdown handler registered in OCA.Viewer for editing (Fixes: #6)
Turns out OCA.Viewer exposes `available_handlers`
for the mime types that are currently supported.

The handlers themselve have a component attribute
that contains the Vue.js component used to render them.

We get the handler for the `text/markdown` mime-type
and use its component to render the editor for the file
we are working on.

This makes it so we can use the same markdown editor
as the files app.

In the case of the text app markdown editor
it even provides us with multi party synchronized editing :)

Also, create the page right away as a file on the server
and update the view accordingly.

This removes the intermediate step of displaying a form
that requires a Save button to be clicked.

Right now there is no way to change the title though.
My idea is to update the title according to the main heading
of the document.