Back-end: rename instead of updating pages
Now that the Text app is responsible for editing the content of our
pages, the back-end no longer needs to be capable of doing so.

This is a first step to remove obsolete logic from the backend: we no
longer allow to update the page content through our back-end. Instead,
we merely rename pages.