Front-end: Add form to rename page (Fixes: #7)
Add a Vue.js computed property `newTitle` which returns the page title
if `currentNewTitle` is empty and the `currentNewTitle` otherwise. This
is necessary in order to separate handling of page title form content
and `this.currentPage.title` as they diverge while writing in the form
and before hitting the "rename" button.

We have to take care of updating `newTitle` everytime `this.currentPage.title`
might have changed. Therefore, the computed property `newTitle` also has
a setter in order to update `this.currentNewTitle`. Unfortunately, defining a
watch property didn't work here.