Front-end: remember new titles without rename
If one changes the title field
and navigates to the next page
and then navigates back
the changed title field is still displayed.

This is implemented by keeping track of a `newTitle`
for every page in the list.

Also moved from a computed property
to a watch expression to initialize `newTitle`
with the current title.