fix: preview changes when changing page
Add a key to the preview component so it gets updated
when the currentPage.id changes.

Keys need to be unique so prefix the editor and the preview keys.

Also cleanup some things that were copied from the NC text app
to the preview but only relate to editing mode.