List file versions in the sidebar (Fixes: #15)
Try to fetch the version history of a file via DAV API, parse the
resulting XML and show a list of old versions in the sidebar.

Copied code from workflowengine app to parse XML output as apparently
there's no exposed functions in Nextcloud to do so yet.

Also copied markup code from files_versions app to list the old versions
in the sidebar.