ui: tweaks to page creation flow
You click "New Page"
* The title input will be focussed and empty with a "Title" placeholder
* the editor will be hidden
* you can open the editor by clicking the toggle

You fill in the title and press enter, or click somewhere else
* The editor becomes visible
* the editor gets focus.

This way it's clear that the app expects you to fill in a title.
If you really want to postpone that there's a way around it.

You cannot just click enter
because an empty title will not be accepted.

fixes #29
fixes #24