routes: use /name.md?fileId=123 format
That is also the format proposed for the nextcloud text app
in https://github.com/nextcloud/text/pull/879.

This commit changes the router-links in the nav.
It also works with links created by NC text
but it will open them in new tabs
because that is what tiptap does by default.