Backend: Add support for wiki folders
Every wiki now has a folder that is supposed to be available to members
of the wiki circle (e.g. by being shared with the circle). Every call to
the `_pages` API now requires a `wikiId` and the scope of actions on
pages is limited to the associated wiki folder.