ui: change icons for Wiki pages to squares with colored shadows
Nextcloud uses colored circles with a single letter
as a placeholder for user icons.

If we use them for wiki pages
that becomes confusing.

Pages are usually rectangular and white with letters.
So use a style that reflects this.

At the same time apply colors so pages are easier to remember.

I tried different options:
* solid borders - make the letters look wrongly aligned
  as the border seems to go towards the inside on the low end.
* outlines - can only be used as squares
  and remind me of MS products too much.
* borders on the left side - would be an alternative.