Use wiki name as circle and folder names
We decided to use the chosen wiki name directly as circle name, without
trying to make it unique (by adding a UUID). This likely will result in
namespace conflicts when creating a wiki, but at the same time it will
prevent namespace conflicts at a later time (e.g. when a user is added
to a wiki circle with a wiki name that conflicts with an existing one).

Also update the README to reflect current state of wiki+circle
implementation.