cleanup: translation tool workarounds
The translationtool used to also consider all directories
in `node_modules` that looked like an app to it.

This is fixed upstream now
so we can remove our workarounds.

Also standardize the paths in the .pot file to simply `/unite/`.
That way commits should only contain the actually changed keys.

fixes #85