ux: pick emoji separately from name
Names with emojis mixed into them is not what we want.

We'd like the emojis in front of the name in the ui
to replace the black star.

We'd also like the emoji at the end of the filename
so we can still type in the names in the command line
and just use auto completion for the emoji.

- [x] keep the emoji separate from the name in the entry
- [x] store the emoji as a postfix of the name
- [ ] pick the emoji from the name before displaying it
- [ ] display the emoji instead of the star

Also includes some small tweaks to the editor display:
* do not cut off emojis and the buttons
* move timestamp closer to check mark button