fix: handle missing href matches gracefully
When the regexp to parse the href returned no matches
for example because the url followed a different patern
match will return null.

In this case the assignment is effectively `[_, id] = null`
and fails because null is not iterable.

Prevent this by wrapping the match in a function
that returns [] instead of null if no match was found.