fix: check for circles app in mount provider
fixes #110
fixes #117

If the circles app is not installed
its API is not available
and we should not call out to the `CollectivesHelper`

Avoid this by checking if the circles app is installed
in the mount point provider
before calling the `CollectivesHelper`.

This is a quick fix to make sure the app does not break the entire server.
Similar issues still exist with the search providers.
It may be better to wrap the `CollectivesHelper`
in a way that checks for the availability of the Circle API.