test: integration tests with cypress
So far just one test for disabling the circles app.
See #110 #117 for the original issues.

On CI build a docker image and test it with cypress
Use wait-on to wait for the nextcloud instance.

Include a fix for browser crashes according to:
https://github.com/cypress-io/cypress/issues/350#issuecomment-503231128

Update according to new before:browser:launch event syntax:
https://docs.cypress.io/guides/references/migration-guide.html#Plugin-Event-before-browser-launch