Put collective mountpoints in subfolder (Fixes: #94, #123)
We now create a subfolder in users home and put all Collectives
mountpoints there.

The subfolder is created when missing (name: localized "Collectives").
If another non-folder object with the same name already exists, we'll
rename that one beforehands.

We decided against implementing a special mountpoint for `/Collectives`
as that would cause too much confusion with direct filesystem access.