js: handle 422 and 409 http responses on `newCollective`
A 409 indicates that the user is already a member of a collective
with that name... so just show it with an info message.
A 422 indicates that something was wrong with the name. Most likely it is used already.
So show that error message as a warning.

Also reworked much of the error handling:
* For http responses in the 400 range
  only show a warning popup
  and do not log them in the console.
* This log and show error popup on 500 response codes
* Add a mechanism to show info popups from the store.
  Requests may sometimes have interesting info in them
  like the 409 response that indicates that a collective already existed.
  We still consider this response successful.
  So we have to handle the info inside the store.
  Expose it via a getter and then watch it in the main view.
  If the info changes display the new info and clear it.