feature: create collective for circle
If a circle of the given name already exists
try to find a circle by that name that has the user as admin
if there is none throw the `CircleAlreadyExists` error as usual.
If there is one we can use that.

Add the info that we used an existing circle to the response.
Pick infos from the response and render them in the browser.