fix: avoid scrolling the entire `main` content
Both the `.app-content-list` as well as the `text-container` were too large.
The `app-content-list` inherits a hight of 100% - 50px.
However we had an additional padding of 40px in the `#app-content-wrapper`.
In order to make sure it actually fits the screen
move the padding into the `app-content-list`.

The `text-container` had a hight of 100% not taking into account the
`#titleform` above which also needs 56px of space.

This also fixes a flaky cypress test
which broke because the `nav-toggle` was scrolled outside the view.

Fixes #145.