fix: load collectives and pages based on url
In the `mounted` function of the view dispatch the `getCollectives` action.
In the action we await the request and then update `state.collectives`.

This triggers the responsive setter function and it will
mark all dependent computed properties as dirty.
Watch `currentCollective` instead of `collectiveParam`.
The former will change when `state.collectives` is filled
whereas the latter will not (on initial page load).

Avoid async / await in the view

There's two ways of dealing with the async nature of the vuex actions:
* write async functions and await the actions
* use the Promises that the actions return,
  modify it with `.then` and `.catch`
  and return it.

We forgot to await the `handleError` function before.
So working with the promises is probably a better idea.

Before we relied on loading the pages from inside the `mounted`
function.
The idea was to await the `getCollectives` action
and then get the pages for the `currentCollective`.
Since we did not await the `handleError` function
the check for `currentCollective` was performed
before the request even returned.