Backend: Implement support for subpages (#65)
The routes for pages API calls changed. Every API call except `index` now
requires requires a `parentId` as argument:
`/_collectives/{collectiveId}/_pages/parent/{parentId}...`

The pages list returned by the API also changed a bit. Every page has a
`parentId` set. For the landing page of the collective, `parentId` is
`0`, for all other pages, it's the id of the parent page. Also, every
page has `pagePath` and `collectivePath` set.

On the filesystem, each folder is expected to have a `Readme.md` (the
index page of this folder). If that file doesn't exist, an empty one is
created on the fly. These index pages have their `title` set to their
folder names (not `Readme`).

Deleting pages that still have subpages throws a `NotPermittedException`.

Whenever a page is moved or deleted, the subfolder in question is parsed
and index pages (`parent/page/Readme.md`) **without** subpages are moved
to their parent folder (`parent/page.md`).