Frontend: add basic support for subpages (#65)
Component `PageList` only lists the `collectivePage` (the landing page),
all other pages are listed recursively in `SubPagesList`. For each page
it is determined whether it contains further subpages.

Indention is intentionally limited to a maximum of three levels. All
further pages are listed with the same indention to prevent huge
unreadable trees.

All levels except the first one are collapsed by default and can be
uncollapsed with a toggle.