fix: occ commmand when circles app is not installed.
The depenency injection of `MountProvider` in `VersionsBackend`
required `Db/CollectiveMapper` which in turn tried to instantiate
`CircleRequest`.

`CircleRequest` is not defined without the circles app installed.

Use the public Circles API rather than the internal `CircleRequest`.
This is slow because it fetches a list of members and linked groups
just to get the name of a collective.

However we can avoid using it in the `MountProvider` which should be the
only performance critical spot.

Move `getMount` from `MountProvider` into `CollectiveFolderManager`.
This way `VersionsBackend` does not need to instantiate a
`MountProvider`.

*Test* occ cli works without circles app

Fixes #167.