fix: make links work in subpages
We use the `Readme.md` file as an index file.
If one opens a Page that has subpages
the actual content lives in the directories `Readme.md` file.
However we only display the Page name and not the `Readme.md`
that would be attached to the directory name.

When resolving links within a collective we need to take this into account.
Visiting `Collective/Container` we are actually on `Collective/Container/Readme`.
So a link to `Collective/OtherPage` will be `../OtherPage`
and a link to `Collective/Container/Sub` will be `Sub`.

If we resolve these links as they are
in the second case the `Container` part will be replaced with `Sub`
as it is the last part of the path.

So we need to prefix `Sub` with `Container`.
`Container/Sub` will then replace `Container`.

Prefixing `../OtherPage` with `Container` also works.
It will effectively result in `OtherPage` with is what we want
relative to `Collective/Container`.

Fixes #35.