fix: show and hide `titlePlaceholder` explicitely
This way the placeholder will not be hidden by `GET_PAGES`
because it is independent of the `pages` state in vuex.

`pages` map the server content more closely this way
and we do not risk loosing state when updating it in the background.

However we do not keep track of pages that were created and whose title has not been changed.
Therefore navigating away from a new page and coming back the title will be "New Page"
and the placeholder will not be shown.

This behavior seems consistent and easy to understand.