ui: button to create collective on `Home`
We do not display the menu by default anymore on mobile.
So give people a hint where they can get started.

Also replace the instruction with the short description of the app

Create a `CollectiveListItem` to use in the `Nav`.

Fixes #88.

rename: `CollectiveList` to `CollectiveListItem`

ui: also show `Create new Collective` button on desktop