ui: focus the title instead of the editor if possible
This requires the text `ViewerComponent` to hand the `autofocus` prop
on to the `EditorWrapper` component.
( https://github.com/nextcloud/text/pull/1645 )

Fall back gracefully for older versions of text.
In particular do not focus the title
as the editor will steal the focus short after
which is really disruptive when people are typing.