fix: duplicate subscription to `start-new-collective`
The `#list` slot of the `AppNavigation` component was rendered twice.
Once before the list of collectives was loaded.
And then again afterwards.

Both times the `NewCollective` component was mounted
but for some reason it was not unmounted in between.

This resulted in two subscriptions to the `start-new-collective` event
one of which did not have a proper component to work with anymore.
This one would complain about a missing `$refs.nameField` when triggered.

Always show the `#list` slot of the navigation.
This way the components are only mounted once
and the list content updates as `collectives` are filled in the store.