fix: keep `occ` working even without `circles` app
When we initialize the `CircleHelper`
catch errors triggered by a missing circles app.

Keep track of the error message and throw
`MissingDepencencyException`
once we attempt to use the `CircleHelper`
in a way that depends on the circles app.

Before the `CircleHelper` used dependency injection
to provide the `CircleManager`.
This triggered exceptions when instantiating the `CircleHelper`
when the circles app was not installed.

`occ` tasks instantiate their dependencies.
Therefore the `CircleHelper` may not throw exceptions
in the constructor.
Otherwise no `occ` command can be run.