Feature: print button for each collective
Open the collective and render all the subpages.
Once they are all rendered open the print dialogue.

Includes a fix to make the `ready` event of `RichText` wait
until the document has actually been rendered.

Inside `Subpages` component keep track of all the pages
and their subpages
that still need to be rendered.
Emit `ready` when this is done.

Inside the `Page` component
wait for the `RichText` and `Subpages` component
to both emit `ready`
and then trigger the print dialogue.