feature: drag page from list to add page link
Make the page list entry generate a meaningful `a` tag when dragged and dropped.

This will in particular make it easy to add links to the page edited right now.

Adding a full url since the page might also be dragged to other programs
such as document editors.
These have no way of making sense of relative paths.