test: wait for page list to load with longer timeout
Loading the pages of a collective in ci seems to be slow.
Allow all dom actions (`get`, `click`, etc) to take 10 seconds.

Make the test more robust by not waiting for `#text h1` to show up on time.

Also remove a `screenshot()` that was left over from a debugging session.