tests: fix flaky tests
Make sure the page content has stabelized so there are no requests pending after the test.

Before the test succeeded even when the sidebar had not been updated yet.
So sometimes the next test would start, clear the cookies
and the old page would still be open and submit the request to update the page list.
This would lead the server to respond with a 401
- because deleting the cookies effectively logged the user out.

Fixes #263.