Store collectives folder for users in database
Instead of getting and translating the collectives user folder each time
the filesystem is initialized, just do that the first time and store the
folder path in database as a user setting.

This fixes issues with several collectives folders (in different
languages) being created in the user home directory. Regardless of
different clients with different languages, we now always use the folder
path that's stored for the user in database.

It also should improve performance, as we no longer need to get the user
language everytime we initialize the filesystem.

Fixes: #238