fix: links between collective pages (regression).
Make use of the `click-link` event from the `ReadOnlyEditor`
to open links to pages inside the collectives app directly.

Port the logic previously provided by the `LinkView`.

Tiptap 2 does not allow for vue views for marks.
Rely on the prosemirror plugin that handles `handleClick`
in the text link mark instead.

Signed-off-by: Max <max@nextcloud.com>