ci: retry tests that fail
See https://docs.cypress.io/guides/guides/test-retries.

Obviously this only helps with tests that can successfully be retried.
If the test creates data and then fails
it may fail in the retry just because the data already exists.

However it seems worth a try.
This way we can try to either make a test reliable
or make it retriable.

Once we have achieved that we can work on making all tests reliable AND retriable.