Always pass pageContent to RichText component
Outsource the `getPageContent()` functions into a mixin.

In Page component, when changing from edit to read mode, take content
from Editor and compare it with the previous content to detect changes.
Also, pass this content to RichText instead of fetching it again there.

In Version and PagePrint components, fetch the content as well and pass
it to RichText.

In RichText, add `updatedPageContent()` method as a hack to enforce
reactivity.