Open non-empty list of collectives on app start on mobile
We don't have much to share on our app start page anyway, so let's
display the list of collectives also on mobile, if there's already
collectives.

Since the list of collectives is fetched asynchronously in the
background, use a watcher in Home.vue to open the navigation once the
list of collectives in store changes from 0 to more.

This brings the nice sideeffect of a slight delay when loading the
navigation - which draws the users attention to it.

Fixes: #341