Improve page drag and drop experience
If a page doesn't have a subpage yet (or it's not expanded), dragging
another page underneath it doesn't create a new subpage list. This
behaviour created too much UX confusion at selecting the target level
when dragging a page at the end of expanded subpages.

Instead, dragging a page *over* a page without subpages will highlight
the dragover page and dropping will turn the dragged page into a subpage
of the dragover page.

For this, in addition to the Sortable.js logic, we add custom drag and
drop handlers in Item.vue.

Also, when a page moving operation is pending (waiting for the backend
response), moving/sorting is disabled now to prevent race conditions.

Related: #423

Signed-off-by: Jonas <jonas@freesources.org>