Fix issue with sortableJS when moving last element from sublist
When moving the last page from a subpage list by dragging it over
another page, the SortableJS ghost element didn't disappear afterwards.

That's because we already removed the `Draggable` element that contained
the SortableJS instance by clearing the subpage list.

Do fix this issue, when moving a page between lists, we now keep the
sortableJS element of the source list in the DOM until the move
operation finished.

Signed-off-by: Jonas <jonas@freesources.org>