Fix handling of deleted pages with subpages
In the filesystem trash backend, the folder gets deleted - and the
fileId of this folder node is added to the page trashbin in DB.

On the other hand, in the collectives pages representation in DB,
we have to make sure that always the page file itself gets referenced.
In case of page with subpages, this is the index file inside the folder
node.

Signed-off-by: Jonas <jonas@freesources.org>