enh(api): js wrapper for collectives api
Wrap all usage of the collectives api in a simple functional interface.
This will allow us to change the layout of the api
and only update it in a single place.

Separate the API handling from the store.
The api is concerned with axios and generateUrl
while the store is only concerned with the current state of the app.

Accept multiple arguments to `collectivesUrl` and `pagesUrl`
and join them with `/` to build the url:

Example:
```js
api.pagesUrl({collectiveId: 10}, 'trash', 12)
```
results in `/apps/collectives/_api/10/_pages/trash/12`

Use the api directly in the cypress commands.
That way we do not require the page to be loaded.

Also extract `collective-edit-mode.spec.js` from `pages.spec.js`.
It requires a different beforeEach function
to seed the edit mode before visiting the collective.

Signed-off-by: Max <max@nextcloud.com>