test(cy): make dispatch available as child command
When used as a child command expects an object to be yielded
and merges it into the payload:
`cy.wrap({ id: 123 }).dispatch(SOME_ACTION, { value: 'Hello' })`
will dispatch `SOME_ACTION` with a payload of `{ id: 123, value: \'Hello\'}`.

If null is yielded the action won't be dispatched.
This is useful for cleanup commands like `deleteCollective`.

Signed-off-by: Max <max@nextcloud.com>