test(cypress): Refactor and adjust tests to open links for Nextcloud 29
Split tests by view/edit mode to see more easily which one fails.

Starting with Nextcloud 29, clicking on a link opens the link bubble.
This bubble contains a preview, clicking on it opens the link.

Also starting with Nextcloud 29, all links to the collectives app open
in the same tab, regardless whether in edit or view mode.

Signed-off-by: Jonas <jonas@freesources.org>