test(cypress): Wait for editor before clicking a link
There might be a race condition between clicking a link in view mode and
resetting the MarkdownContentEditor (view mode) content when the Editor
(edit mode) finished loading.

Therefore let's wait for the editor before clicking on links.

Also, remove a superfluous extra click on links in view mode that was
needed as a workaround earlier.

Signed-off-by: Jonas <jonas@freesources.org>