test(ci): Run psalm against app store release with latest migrations
We had upgrade failures already a few times in the past because a
migration or repair step referenced a new constant or function in
an existing class.

Due to the way app upgrades are handled in Nextcloud, the same PHP
process that patches the app sources with the new ones from the app
store then runs the new migrations and repair steps. Existing classes
don't get refreshed before that, so depending on new objects in an
existing PHP class can cause problems here.

This commit adds a CI check that statically checks the migrations and
repair steps from latest development branch along with the app code of
the former app store release.

This should catch situations where we accidentally depend on new
objects in migrations.

Signed-off-by: Jonas <jonas@freesources.org>