fix(notifyPush): only update the current collective
When switching to a different collective
we reload the page list anyway.

We only see updates to other collectives if there is a session open
so the collective is open in some other tab - or maybe even some other device.

If it is another tab we also receive the updates via localStorage.

But right now page lists of other collectives are just an approximation
anyway until the collective is opened and the list updated.

Signed-off-by: Max <max@nextcloud.com>