feat(serviceWorker): Custom fallback mechanism for vue-router URLs
Instead of trying to use `navigateFallback` which relies on a static
HTML file as entry point, use a custom fallback mechanism with a single
cache key for all URLs that are routed in the frontend anyway.

The workbox runtime caching plugins sets one single cache key for all
URLs that are routed in the frontend and in case of network issues the
cache entry will be servced for all those URLs.

Signed-off-by: Jonas <jonas@freesources.org>