fix(load): Avoid loading pagelist flipflop
It was initialized as false,
set to true when `CollectivesApp` was mounted
and then set to false again after loading the page list.

This led to the TextEditor being loaded twice.

**Alternatives considered**

We'll probably want a different logic
that takes into account the page list already present in local storage.

But that's for another time.

Signed-off-by: Max <max@nextcloud.com>