fix(EmojiPicker): Use event-bus instead of store flag to open
The `showings` object in root store is persisted in localStorage and we
don't want the emoji picker state to be persisted.

Also it eems cleaner to use an event-bus event for doing this, as we
only want a trigger to *open* the emoji picker. We don't want to *close*
it programmatically anyway.

Signed-off-by: Jonas <jonas@freesources.org>