fix(store): Separate collectives and pages store for public shares
In collectives store, use separate states for internal and public
collectives. `publicCollectivesState` is an object with `shareToken`
as index now, `collectivesState` stays a simple array.

In pages store, the state objects `allPages`, `allTrashPages` and
`allAttachments` now keep internal collectives with `collectiveId`
as index and public collectives with `shareToken` as index.

This should fix running into conflicts when opening subpage shares in
same browser context as the authenticated collective.

Signed-off-by: Jonas <jonas@freesources.org>